+++ mapreduce/src/list.h
@@ -24,7 +24,7 @@ typedef list_node* LNode;
 /* List control node */
 typedef struct _list_root {
   LNode root_node; /* The first list node */
-  unsigned long len; /* length of list */
+  unsigned long long len; /* length of list */
 } list_root;
 typedef list_root* List;
 
@@ -41,7 +41,7 @@ int list_node_destroy_hard(LNode l, int (*destroyer) () );
 int LPush(List l, list_data_t value);
 list_data_t LPop(List l);
 LNode LSearch(List l, list_data_t value);
-list_data_t LAt(List l, unsigned long ind);
+list_data_t LAt(List l, unsigned long long ind);
 
 /* Reverse the list */
 int LReverse(List l);
@@ -50,10 +50,10 @@ int LReverse(List l);
 int list_node_push(LNode* l, list_data_t value);
 list_data_t list_node_pop(LNode* l);
 LNode list_node_search(LNode l, list_data_t value);
-int list_node_assign(LNode l, list_data_t* values, const unsigned long values_len);
+int list_node_assign(LNode l, list_data_t* values, const unsigned long long values_len);
 
 /* Some more utils */
-unsigned long LLen(List l);
+unsigned long long LLen(List l);
 int LCpy(List l, const List o);
 /* converts some array to List
    Warning!! the source array will be destroyed
@@ -62,7 +62,7 @@ int LCpy(List l, const List o);
 List AtoL(void* some_array[], unsigned long long arr_len);
 
 
-unsigned long list_node_len(LNode l);
+unsigned long long list_node_len(LNode l);
 int list_node_find(LNode l, list_data_t value);
 int list_node_find_root(LNode* l);
 int list_node_delete_node(LNode l, list_data_t value);
